requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
sqlalchemy==2.0.48
alembic==1.18.4
psycopg2-binary==2.9.11
python-dotenv==1.2.2
pydantic[email]==2.12.5
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
